# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiError": {
        "description": "Forma única de todos los errores. La UI la traduce a un mensaje por `code`.",
        "properties": {
          "code": {
            "description": "Identificador estable del error, p. ej. NotFoundError",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Datos extra del error",
            "title": "Detail"
          },
          "message": {
            "description": "Mensaje en español, ya legible",
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "CreateUserIn": {
        "properties": {
          "email": {
            "minLength": 3,
            "title": "Email",
            "type": "string"
          },
          "mustChangePassword": {
            "default": true,
            "title": "Mustchangepassword",
            "type": "boolean"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          },
          "password": {
            "minLength": 12,
            "title": "Password",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        },
        "required": [
          "email",
          "role",
          "password"
        ],
        "title": "CreateUserIn",
        "type": "object"
      },
...
      "Role": {
        "description": "Roles del equipo. Se corresponden con los del prototipo (`governance.ts`).",
        "enum": [
          "administrador",
          "lider_tecnico",
          "qa",
          "desarrollador",
          "arquitecto",
          "analista",
          "observador"
        ],
        "title": "Role",
        "type": "string"
      },
...
      "UpdateUserIn": {
        "properties": {
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Role"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "UpdateUserIn",
        "type": "object"
      },
...
      "UserOut": {
        "description": "Una persona. **Sin contraseña ni hash**: `User` del dominio tampoco los tiene.",
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "initials": {
            "title": "Initials",
            "type": "string"
          },
          "lastLoginAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastloginat"
          },
          "locked": {
            "default": false,
            "title": "Locked",
            "type": "boolean"
          },
          "mustChangePassword": {
            "default": false,
            "title": "Mustchangepassword",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "permissions": {
            "items": {
              "type": "string"
            },
            "title": "Permissions",
            "type": "array"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "roleLabel": {
            "title": "Rolelabel",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "name",
          "initials",
          "role",
          "roleLabel",
          "provider",
          "active",
          "permissions",
          "createdAt"
        ],
        "title": "UserOut",
        "type": "object"
      },
      "UserPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/UserOut"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "UserPage",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Backend del orquestador multiagente. El dominio gira alrededor de un Requerimiento: adjuntos, plan de agentes, ejecuciones y entregables.",
    "title": "Orquestador de revisión de requerimientos",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/users": {
      "get": {
        "operationId": "index_api_users_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Petición inválida"
          },
          "402": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Presupuesto de la ejecución agotado"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Prohibido por política (PHI, BAA, herramienta)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "No existe"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflicto: el plan está bloqueado"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "La petición no tiene el formato esperado"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error interno"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Fallo del proveedor de IA"
          }
        },
        "summary": "Index",
        "tags": [
          "identidad"
        ]
      },
      "post": {
        "operationId": "create_api_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Petición inválida"
          },
          "402": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Presupuesto de la ejecución agotado"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Prohibido por política (PHI, BAA, herramienta)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "No existe"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflicto: el plan está bloqueado"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "La petición no tiene el formato esperado"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error interno"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Fallo del proveedor de IA"
          }
        },
        "summary": "Create",
        "tags": [
          "identidad"
        ]
      }
    },
    "/api/users/{user_id}": {
      "put": {
        "description": "Cambia rol, nombre o si la cuenta está activa.\n\nDesactivar o cambiar de rol **cierra las sesiones abiertas** de esa persona: si no, el cambio\nno surtiría efecto hasta que cerrara sesión por su cuenta.",
        "operationId": "update_api_users__user_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Petición inválida"
          },
          "402": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Presupuesto de la ejecución agotado"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Prohibido por política (PHI, BAA, herramienta)"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "No existe"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Conflicto: el plan está bloqueado"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "La petición no tiene el formato esperado"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Error interno"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Fallo del proveedor de IA"
          }
        },
        "summary": "Update",
        "tags": [
          "identidad"
        ]
      }
    }
  }
}